feat: remove old migrations
