+++ internal/initialization/initialization.go
@@ -9,13 +9,10 @@ import (
 	"github.com/mini-maxit/backend/internal/api/queue"
 	"github.com/mini-maxit/backend/internal/config"
 	"github.com/mini-maxit/backend/internal/database"
-	"github.com/mini-maxit/backend/package/domain/schemas"
-	"github.com/mini-maxit/backend/package/domain/types"
 	"github.com/mini-maxit/backend/package/filestorage"
 	"github.com/mini-maxit/backend/package/repository"
 	"github.com/mini-maxit/backend/package/service"
 	"github.com/mini-maxit/backend/package/utils"
-	"go.uber.org/zap"
 
 	amqp "github.com/rabbitmq/amqp091-go"
 )
@@ -76,64 +73,19 @@ func NewInitialization(cfg *config.Config) *Initialization {
 	if err != nil {
 		log.Panicf("Failed to connect to database: %s", err.Error())
 	}
-	tx, err := db.BeginTransaction()
 
-	if err != nil {
-		log.Panicf("Failed to connect to database: %s", err.Error())
-	}
 	// Repositories
-	langRepository, err := repository.NewLanguageRepository(tx)
-	if err != nil {
-		log.Panicf("Failed to create language repository: %s", err.Error())
-	}
-	userRepository, err := repository.NewUserRepository(tx)
-	if err != nil {
-		log.Panicf("Failed to create user repository: %s", err.Error())
-	}
-	fileRepository, err := repository.NewFileRepository(tx)
-	if err != nil {
-		log.Panicf("Failed to create file repository: %s", err.Error())
-	}
-	taskRepository, err := repository.NewTaskRepository(tx)
-	if err != nil {
-		log.Panicf("Failed to create task repository: %s", err.Error())
-	}
-	groupRepository, err := repository.NewGroupRepository(tx)
-	if err != nil {
-		log.Panicf("Failed to create group repository: %s", err.Error())
-	}
-	submissionRepository, err := repository.NewSubmissionRepository(tx)
-	if err != nil {
-		log.Panicf("Failed to create submission repository: %s", err.Error())
-	}
-	_, err = repository.NewSubmissionResultRepository(tx)
-	if err != nil {
-		log.Panicf("Failed to create submission result repository: %s", err.Error())
-	}
-	testCaseRepository, err := repository.NewTestCaseRepository(tx)
-	if err != nil {
-		log.Panicf("Failed to create input output repository: %s", err.Error())
-	}
-	testResultRepository, err := repository.NewTestResultRepository(tx)
-	if err != nil {
-		log.Panicf("Failed to create test result repository: %s", err.Error())
-	}
-	queueRepository, err := repository.NewQueueMessageRepository(tx)
-	if err != nil {
-		log.Panicf("Failed to create queue repository: %s", err.Error())
-	}
-	submissionResultRepository, err := repository.NewSubmissionResultRepository(tx)
-	if err != nil {
-		log.Panicf("Failed to create submission result repository: %s", err.Error())
-	}
-	contestRepository, err := repository.NewContestRepository(tx)
-	if err != nil {
-		log.Panicf("Failed to create contest repository: %s", err.Error())
-	}
-
-	if err := db.Commit(); err != nil {
-		log.Panicf("Failed to commit transaction: %s", err.Error())
-	}
+	langRepository := repository.NewLanguageRepository()
+	userRepository := repository.NewUserRepository()
+	fileRepository := repository.NewFileRepository()
+	taskRepository := repository.NewTaskRepository()
+	groupRepository := repository.NewGroupRepository()
+	submissionRepository := repository.NewSubmissionRepository()
+	testCaseRepository := repository.NewTestCaseRepository()
+	testResultRepository := repository.NewTestResultRepository()
+	queueRepository := repository.NewQueueMessageRepository()
+	submissionResultRepository := repository.NewSubmissionResultRepository()
+	contestRepository := repository.NewContestRepository()
 
 	// Services
 	filestorage, err := filestorage.NewFileStorageService(cfg.FileStorageURL)
@@ -228,72 +180,5 @@ func NewInitialization(cfg *config.Config) *Initialization {
 		WorkerRoute:     workerRoute,
 
 		QueueListener: queueListener,
-	}
-}
-
-func dump(db *database.PostgresDB, log *zap.SugaredLogger, authService service.AuthService, userRepository repository.UserRepository) {
-	tx, err := db.BeginTransaction()
-	if err != nil {
-		log.Warnf("Failed to connect to database to init dump: %s", err.Error())
-	}
-	users := []struct {
-		Name     string
-		Surname  string
-		Email    string
-		Username string
-		Password string
-		Role     types.UserRole
-	}{
-
-		{
-			Name:     "AdminName",
-			Surname:  "AdminSurname",
-			Email:    "admin@admin.com",
-			Username: "admin",
-			Password: "adminadmin",
-			Role:     types.UserRoleAdmin,
-		},
-		{
-			Name:     "TeacherName",
-			Surname:  "TeacherSurname",
-			Email:    "teacher@teacher.com",
-			Username: "teacher",
-			Password: "teacherteacher",
-			Role:     types.UserRoleTeacher,
-		},
-		{
-			Name:     "StudentName",
-			Surname:  "StudentSurname",
-			Email:    "student@student.com",
-			Username: "student",
-			Password: "studentstudent",
-			Role:     types.UserRoleStudent,
-		},
-	}
-	for _, user := range users {
-		_, err = authService.Register(tx, schemas.UserRegisterRequest{
-			Name:     user.Name,
-			Surname:  user.Surname,
-			Email:    user.Email,
-			Username: user.Username,
-			Password: user.Password,
-		})
-		if err != nil {
-			log.Warnf("Failed to create %s: %s", user.Username, err.Error())
-		}
-		registeredUser, err := userRepository.GetByEmail(tx, user.Email)
-		if err != nil {
-			log.Warnf("Failed to get %s user: %s", user.Username, err.Error())
-		}
-		registeredUser.Role = user.Role
-		err = userRepository.Edit(tx, registeredUser)
-		if err != nil {
-			log.Warnf("Failed to set %s role: %s", user.Username, err.Error())
-		}
-	}
-
-	err = db.Commit()
-	if err != nil {
-		log.Warnf("Failed to commit transaction after dump: %s", err.Error())
 	}
 }
