+++ package/repository/file.go
@@ -20,16 +20,6 @@ func (f *file) Create(tx *gorm.DB, file *models.File) error {
 	return nil
 }
 
-func NewFileRepository(db *gorm.DB) (File, error) {
-	tables := []any{&models.File{}}
-	for _, table := range tables {
-		if !db.Migrator().HasTable(table) {
-			err := db.Migrator().CreateTable(table)
-			if err != nil {
-				return nil, err
-			}
-		}
-	}
-
-	return &file{}, nil
+func NewFileRepository() File {
+	return &file{}
 }
